+++ index.js
@@ -1,6 +1,6 @@
-const  mdLinks = require('./hito1.js');
+const  mdLinks = require('./md-links.js');
 
-// mdLinks('./fileMD/archivo.txt')
+//mdLinks('./fileMD/archivo.txt')
 mdLinks('./fileMD/links.md',true)
 .then(links => {
   console.log(links);
@@ -9,6 +9,12 @@ mdLinks('./fileMD/links.md',true)
   console.error(error.message);
 });
 
-
+/*mdLinks('./fileMD/', true, (error, result) => {
+  if (error) {
+    console.error('Error:', error);
+  } else {
+    console.log('Resultado:', result); 
+  }
+});*/
 // llamamos la funcion de md-links
